+++ core/load/Loadable.ts
@@ -4,6 +4,7 @@
  *
  * Copyright © 2026 Extremely Heavy Industries Inc.
  */
+import {SpanConfig} from '@xh/hoist/utils/telemetry';
 import {TaskObserver} from '../';
 import {LoadSpec} from './';
 
@@ -25,6 +26,13 @@ export interface Loadable {
 
     /** Any exception that occurred during last load. */
     lastLoadException: any;
+
+    /**
+     * Set to auto-generate spans covering the loads on this object.
+     * The managed span will be placed on the LoadSpec passed to doLoadAsync and available for
+     * tagging, sub-spanning, and passing through to other models and fetch calls.
+     */
+    loadSpan?: Omit<SpanConfig, 'parent'> | string;
 
     /**
      * Load the target.